pyproject.toml
[tool.poetry]
name = "wc"
version = "0.1.0"
description = ""
authors = ["Brendan Cazier <520246+cazier@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
peewee = "^3.16.0"
pyyaml = "^6.0"
typer = "^0.7.0"
rich = "^13.3.5"
faker = "^18.9.0"
ipython = "^8.13.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ipython = "^8.12.0"
ruff = "^0.0.261"
pre-commit = "^3.2.2"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
length_sort = true
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
